requirements.txt
Django>=1.6.2
Jinja2>=2.7.1
hiredis>=0.1.1
passlib>=1.6.2
python-dateutil>=2.2
pytz>=2014.1
redis>=2.9.1
sockjs-tornado>=1.0.0
tornado>=3.1.1
tornado-redis>=2.4.14
git+git://github.com/donaldharvey/voluptuous
